+++ packages/ui/src/components/select/select.component.tsx
@@ -1,24 +1,19 @@
 'use client';
 
 import React, { ForwardedRef, forwardRef } from 'react';
-import { mergeProps, useFocusRing } from 'react-aria';
 
-import { styles } from './select.styles.js';
+import { SelectWithTooltip, StyledSelect } from './components/index.js';
 import { type SelectProps } from './select.types.js';
 
-function BaseSelect(
-  { className, size = 'medium', invalid = false, width = 'auto', children, ...props }: SelectProps,
-  ref: ForwardedRef<HTMLSelectElement>,
-) {
-  const { isFocused, isFocusVisible, focusProps } = useFocusRing();
-  return (
-    <select
-      ref={ref}
-      className={styles({ className, size, invalid, isFocused, isFocusVisible, width })}
-      {...mergeProps(props, focusProps)}
-    >
+function BaseSelect({ children, enableTooltip, ...props }: SelectProps, ref: ForwardedRef<HTMLSelectElement>) {
+  return enableTooltip ? (
+    <SelectWithTooltip ref={ref} {...props}>
       {children}
-    </select>
+    </SelectWithTooltip>
+  ) : (
+    <StyledSelect ref={ref} {...props}>
+      {children}
+    </StyledSelect>
   );
 }