+++ packages/ui/src/components/tooltip/components/tooltip-content/tooltip-content.types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export type TooltipContentProps = {
+  children?: ReactNode;
+  id: string;
+};